Attempt to fix compilation warnings

+++ ElementsExamples/src/program/OpenMPExample.cpp
@@ -70,7 +70,7 @@ public:
         char c2 = ' ';
         if (n > 0) {
           static const char charset[] = ".,c8M@jawrpogOQEPGJ";
-          c2 = charset[n % (sizeof(charset)-1)];
+          c2 = charset[static_cast<std::size_t>(n % (sizeof(charset)-1))];
         }
         std::putchar(c2);
           if (x+1 == width) {
